+++ packages/syllst-processor/src/loader/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @module @syllst/processor/loader
+ * Content loader factory for syllabus content
+ */
+
+export * from './factory.js';