+++ packages/syllst/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @syllst/core
 
+## 0.5.2
+
+### Patch Changes
+
+- Patch version bump for republishing
+
+## 0.5.1
+
+### Patch Changes
+
+- Patch version bump for republishing
+
 ## 0.4.2
 
 ### Patch Changes